extract duplicated logic
some parts for the property_export are
as well in the abstract export

this is bad but better than before :-)
